mod installer

+++ install.sh
@@ -22,7 +22,33 @@ for file in .??*; do
   [[ "${file}" == ".DS_Store" ]] && continue
   [[ "${file}" == ".github" ]] && continue
   [[ "${file}" == ".vscode" ]] && continue
-  ln -fvns "${DOTPATH}/${file}" "${HOME}/${file}"
+  if [ -e "$file" ]; then
+      echo "$file is existing. Skip."
+      continue;
+  else
+      echo "Creating symlink for $file"
+      ln -fvns "${DOTPATH}/${file}" "${HOME}/${file}"
+  fi
 done
+
+if [ ! -d "$HOME/.config/nvim" ]; then
+    echo "Creating ~/.config/nvim"
+    mkdir -p "$HOME/.config/nvim"
+fi
+target=${DOTPATH}/.config/nvim/init.vim
+echo "Creating symlink for $target"
+ln -fvns "$target" "${HOME}/.config/nvim/init.vim"
+
+# Install plugins managers
+## Install zplug
+if [[ ! -d ~/.zplug ]];then
+  git clone https://github.com/zplug/zplug ~/.zplug
+fi
+
+## Install vim-plug
+if [ ! -e  "${XDG_DATA_HOME:-$HOME/.local/share}/nvim/site/autoload/plug.vim" ]; then
+    sh -c 'curl -fLo "${XDG_DATA_HOME:-$HOME/.local/share}"/nvim/site/autoload/plug.vim --create-dirs \
+           https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim'
+fi
 
 echo "Setup done."
